+++ lib/docxify/element/numbered_list_item.rb
@@ -1,9 +1,0 @@
-module DocXify
-  module Element
-    class NumberedListItem < Base
-      def to_s(_container = nil)
-        # TODO: Implement output
-      end
-    end
-  end
-end